Adding French
If someone can explain to me if the accents have to be pre-encoded in UTF-8 in JSON ? If so they will not appear correctly here and the dev will have to encode it after recovering the file. 
You can also add in the file what is needed for the accents to appear correctly.