Patch 1 (#40)
* Adding French

If someone can explain to me if the accents have to be pre-encoded in UTF-8 in JSON ? If so they will not appear correctly here and the dev will have to encode it after recovering the file.
You can also add in the file what is needed for the accents to appear correctly.

* Correcting errors in French

* Update language file for French.

Co-authored-by: Yuan ZHOU <yuanzhou.cq@gmail.com>
(cherry picked from commit 5b425411450ca560a18db5da4eca6b53958e516e)